clean up js tests
